pyproject.toml
[project]
name = "snap-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=24.1.0",
    "asyncpg>=0.30.0",
    "email-validator>=2.3.0",
    "fastapi>=0.116.1",
    "openai>=1.107.3",
    "passlib[argon2]>=1.7.4",
    "pydantic-settings>=2.10.1",
    "pyjwt>=2.10.1",
    "python-multipart>=0.0.20",
    "redis>=6.4.0",
    "rq>=2.6.0",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.35.0",
]


[dependency-groups]
dev = [
    "aiosqlite>=0.21.0",
    "httpx>=0.28.1",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "pytest-mock>=3.14.1",
    "ruff>=0.13.0",
]
